requirements.txt
# Core Requirements
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==3.0.2  # Compatible with Flask 3.0+

# Google AI Integration
google-generativeai==0.3.2  # Version compatible with protobuf<4
protobuf==3.20.3  # Pinned version to prevent conflicts

# Database & JSON Handling
sqlalchemy==2.0.29
python-dotenv==1.0.1

# Image Processing (if needed)
pillow==10.3.0

# Development Server
waitress==2.1.2  # Production server instead of flask's dev server